move ReportView from helpers to controller
